데이터 유효성에 대한 기능을 분리하는 작업
