+++ README.md
@@ -28,3 +28,4 @@
 ### 추가 구현 기능
 > *앞으로 작성 예정*
 * 로또의 6개 번호가 중복되면 안되는 것을 늦게 알아서, 중복 검사 기능 추가
+* 데이터 유효성만 하는 기능을 분리하여 구현
